+++ redistribution/src/blockchain.rs
@@ -46,8 +46,8 @@ impl Blockchain {
                 let previous_block = self.get_latest_block()?;
                 let new_block_index = previous_block.index + 1;
                 let hash = hasher::calculate_hash(&new_block_index, &previous_block.hash, &timestamp, block_data);
-                let difficulty: u128 = 0;
-                let nonce: u32 = 0;
+                let difficulty: u32 = 0;
+                let nonce: u128 = 0;
                 Ok(Block::new(new_block_index, timestamp, block_data.to_string(), hash, previous_block.hash.clone(), difficulty, nonce))
             },
             Err(_) => Err(Error::new(ErrorKind::NotFound, "Error reporting system time"))
